+++ docs/source/documentation.rst
@@ -14,7 +14,7 @@ Key Modules
 
    modules/load_sample_dataset
 
-   modules/output
+   modules/daddy
 
 
 Additional Modules
