+++ docs/source/modules/daddy.rst
@@ -1,0 +1,4 @@
+daddy
+-----
+.. automodule:: pydaddy.daddy
+   :members:
