--- docs/source/modules/output.rst
@@ -1,4 +1,0 @@
-output
-------
-.. automodule:: pydaddy.output
-   :members:
